change precision to int16

+++ subtrPrecision.m
@@ -1,3 +1,3 @@
 function out = subtrPrecision(a,b)
-    out = abs(int8(a) - int8(b)); %works best with int32 precision 
+    out = abs(int16(a) - int16(b)); %works best with int32 precision 
 end
